requirements.txt
openai
numpy
pandas
tqdm
python-dotenv
rank-bm25